Refactor: DogCelebrityMapping 수정
